+++ src/general/cell.h
@@ -30,7 +30,7 @@ protected:
   double measure_value;
 
 public:
-  long int cellno,nodeno[4],faceno[4],neighbour_cellno[4];
+  int cellno,nodeno[4],faceno[4],neighbour_cellno[4];
   NodeType* node[4];
   FaceType* face[4];
